WSHUB-34: ntp: Add broadcast client support.
OpenThread's `wpantund` has some bugs relating to multicast routing that
prevent it from effectively listening to a multicast group.  However, it
can reliably send to a multicast group, and the clients can listen.

This theoretically allows us to broadcast the time from an NTP server
operating on the WPAN to the nodes.  They passively listen for the
broadcasts and update their local clocks accordingly.

Yes, there'll be differences in time, but they should be small.